+++ R/ui_priors.R
@@ -1,6 +1,84 @@
 ### * All functions in this file are exported
 
-### * constant(value)
+### * available_priors()
+
+#' List the available priors for model parameters
+#'
+#' @return A tibble containing information about the available priors.
+#'
+#' @examples
+#' available_priors()
+#'
+#' @export
+
+available_priors <- function() {
+    fun_names <- c("constant_p",
+                   "uniform_p",
+                   "normal_p",
+                   "hcauchy_p",
+                   "exponential_p",
+                   "gamma_p",
+                   "scaled_beta_p")
+    prior_names <- c("Constant value",
+                     "Uniform prior",
+                     "Normal distribution",
+                     "Half-Cauchy distribution",
+                     "Exponential distribution",
+                     "Gamma distribution",
+                     "Scaled beta distribution")
+    usage <- c("constant_p(value)",
+               "uniform_p(min, max)",
+               "normal_p(mean, sd)",
+               "hcauchy_p(scale)",
+               "exponential_p(lambda)",
+               "gamma_p(alpha, beta)",
+               "scaled_beta_p(alpha, beta, scale = 1)")
+    out <- tibble::tibble(function_name = fun_names,
+                          description = prior_names,
+                          usage = usage)
+    return(structure(out, class = c("prior_tibble", class(out))))
+}
+
+### * Methods for nice display of prior tibble
+
+### ** format.prior_tibble()
+
+#' Pretty formatting of a \code{prior_tibble} object
+#'
+#' @param x An object of class \code{prior_tibble}.
+#' @param ... Not used.
+#'
+#' @return A character string for pretty printing of a prior tibble.
+#' 
+#' @export
+
+format.prior_tibble <- function(x, ...) {
+    x <- structure(x, class = class(x)[class(x) != "prior_tibble"])
+    f <- format(x)
+    f <- c(f,
+           "-------------------------------------------------------",
+           "(Note: All priors distributions are truncated at zero.)")
+    return(f)
+}
+
+### ** print.prior_tibble()
+
+#' Pretty printing of a \code{prior_tibble} object
+#'
+#' @param x An object of class \code{prior_tibble}.
+#' @param ... Not used.
+#'
+#' @return Mostly called for its side effect of printing, but also returns its
+#'     input invisibly.
+#' 
+#' @export
+
+print.prior_tibble <- function(x, ...) {
+    cat(format(x), sep = "\n")
+    invisible(x)
+}
+
+### * constant_p(value)
 
 #' Define a fixed-value prior
 #'
@@ -11,18 +89,18 @@
 #' @return A list defining the prior.
 #'
 #' @examples
-#' constant(2)
+#' constant_p(2)
 #'
 #' @export
 
-constant <- function(value) {
+constant_p <- function(value) {
     x <- list(type = "constant",
               parameters = c(value = value))
     x <- structure(x, class = "prior")
     return(x)
 } 
 
-### * hcauchy(scale)
+### * hcauchy_p(scale)
 
 #' Define a half-Cauchy prior (on [0;+Inf])
 #'
@@ -33,18 +111,18 @@ constant <- function(value) {
 #' @importFrom stats rcauchy
 #'
 #' @examples
-#' hcauchy(scale = 0.5)
+#' hcauchy_p(scale = 0.5)
 #'
 #' @export
 
-hcauchy <- function(scale) {
+hcauchy_p <- function(scale) {
     x <- list(type = "hcauchy",
               parameters = c(scale = scale))
     x <- structure(x, class = "prior")
     return(x)
 }
 
-### * normal(mean, sd)
+### * normal_p(mean, sd)
 
 #' Define a truncated normal prior (on [0;+Inf])
 #'
@@ -56,11 +134,11 @@ hcauchy <- function(scale) {
 #' @importFrom stats rnorm
 #'
 #' @examples
-#' normal(mean = 0, sd = 4)
+#' normal_p(mean = 0, sd = 4)
 #'
 #' @export
 
-normal <- function(mean, sd) {
+normal_p <- function(mean, sd) {
     x <- list(type = "trun_normal",
               parameters = c(mean = mean,
                              sd = sd))
@@ -68,7 +146,7 @@ normal <- function(mean, sd) {
     return(x)
 }
 
-### * uniform(min, max)
+### * uniform_p(min, max)
 
 #' Define a uniform prior
 #'
@@ -79,11 +157,11 @@ normal <- function(mean, sd) {
 #' @importFrom stats runif
 #' 
 #' @examples
-#' uniform(min = 0, max= 1)
+#' uniform_p(min = 0, max= 1)
 #'
 #' @export
 
-uniform <- function(min, max) {
+uniform_p <- function(min, max) {
     x <- list(type = "uniform",
               parameters = c(min = min, max = max))
     x <- structure(x, class = "prior")
@@ -91,7 +169,7 @@ uniform <- function(min, max) {
 }
 
 
-### * scaled_beta(alpha, beta, scale=1)
+### * scaled_beta_p(alpha, beta, scale=1)
 
 #' Define a beta prior (on [0;scale])
 #'
@@ -106,13 +184,61 @@ uniform <- function(min, max) {
 #' @return A list defining the prior.
 #'
 #' @examples
-#' scaled_beta(0.8, 20, scale = 10)
+#' scaled_beta_p(0.8, 20, scale = 10)
 #'
 #' @export
 
-scaled_beta <- function(alpha, beta, scale = 1) {
+scaled_beta_p <- function(alpha, beta, scale = 1) {
     x <- list(type = "scaled_beta",
               parameters = c(alpha = alpha, beta = beta, scale = scale))
+    x <- structure(x, class = "prior")
+    return(x)
+}
+
+### * exponential_p(lambda)
+
+#' Define an exponential prior
+#'
+#' @param lambda Lambda parameter (rate) of the exponential distribution. The
+#'     mean of the exponential distribution is 1/lambda.
+#'
+#' @return A list defining the prior.
+#'
+#' @examples
+#' exponential_p(0.5)
+#'
+#' @export
+
+exponential_p <- function(lambda) {
+    x <- list(type = "exponential",
+              parameters = c(lambda = lambda))
+    x <- structure(x, class = "prior")
+    return(x)
+}
+
+### * gamma_p(alpha, beta)
+
+#' Define a gamma prior
+#'
+#' Note the name of the function to define a prior (\code{gamma_p}), in order
+#' to avoid confusion with the R mathematical function \code{gamma}.
+#'
+#' @param alpha Shape parameter (equivalent to the \code{shape} parameter of
+#'     R's \code{rgamma}).
+#' @param beta Rate parameter (equivalent to the \code{rate} parameter of R's
+#'     \code{rgamma}).
+#'
+#' @return A list defining the prior.
+#'
+#' @examples
+#' gamma_p(9, 2)
+#' hist(sample_from_prior(gamma_p(9, 2), 1e3))
+#' 
+#' @export
+
+gamma_p <- function(alpha, beta) {
+    x <- list(type = "gamma",
+              parameters = c(alpha = alpha, beta = beta))
     x <- structure(x, class = "prior")
     return(x)
 }
@@ -135,7 +261,7 @@ format.prior <- function(x, ...) {
                      paste(paste(names(x[["parameters"]]), x[["parameters"]],
                                  sep = "="), collapse = ","),
                      ")")
-    out <- paste0(x[["type"]], " ", params)
+    out <- paste0(x[["type"]], "", params)
     return(out)
 }
 
@@ -210,12 +336,12 @@ pillar_shaft.prior <- function(x, ...) {
 #' @return Boolean (or throws an error for unsupported operators).
 #' 
 #' @examples
-#' p <- constant(0)
-#' q <- constant(4)
+#' p <- constant_p(0)
+#' q <- constant_p(4)
 #' p == q
 #'
-#' p <- hcauchy(2)
-#' q <- hcauchy(2)
+#' p <- hcauchy_p(2)
+#' q <- hcauchy_p(2)
 #' p == q
 #'
 #' @method Ops prior
@@ -248,12 +374,12 @@ Ops.prior <- function(e1, e2) {
 #' @return A numeric vector of length \code{n}.
 #' 
 #' @examples
-#' sample_from_prior(constant(1))
-#' sample_from_prior(constant(1), 10)
-#' sample_from_prior(hcauchy(0.5), 1)
-#' hist(sample_from_prior(hcauchy(0.5), 20))
-#' hist(sample_from_prior(uniform(0, 3), 1000))
-#' hist(sample_from_prior(scaled_beta(3, 7, 2), 1e4))
+#' sample_from_prior(constant_p(1))
+#' sample_from_prior(constant_p(1), 10)
+#' sample_from_prior(hcauchy_p(0.5), 1)
+#' hist(sample_from_prior(hcauchy_p(0.5), 20))
+#' hist(sample_from_prior(uniform_p(0, 3), 1000))
+#' hist(sample_from_prior(scaled_beta_p(3, 7, 2), 1e4))
 #' 
 #' @export
 #'
@@ -267,9 +393,9 @@ sample_from_prior <- function(x, n = 1) {
                scale <- x$parameters[["scale"]]
                replicate(n,
                {
-                   o <- rcauchy(1, location = 0, scale = scale)
+                   o <- stats::rcauchy(1, location = 0, scale = scale)
                    while (o < 0) {
-                       o <- rcauchy(1, location = 0, scale = scale)
+                       o <- stats::rcauchy(1, location = 0, scale = scale)
                    }
                    return(o)
                })
@@ -278,19 +404,28 @@ sample_from_prior <- function(x, n = 1) {
                mean <- x$parameters[["mean"]]
                sd <- x$parameters[["sd"]]
                replicate(n, {
-                   o <- rnorm(1, mean = mean, sd = sd)
+                   o <- stats::rnorm(1, mean = mean, sd = sd)
                    while (o < 0) {
-                       o <- rnorm(1, mean = mean, sd = sd)
+                       o <- stats::rnorm(1, mean = mean, sd = sd)
                    }
                    return(o)
                })
            },
            "uniform" = {
-               runif(n, min = x$parameters[["min"]], max = x$parameters[["max"]])
+               stats::runif(n, min = x$parameters[["min"]], max = x$parameters[["max"]])
            },
            "scaled_beta" = {
                p <- x$parameters
-               p[["scale"]] * rbeta(n, shape1 = p[["alpha"]], shape2 = p[["beta"]])
+               p[["scale"]] * stats::rbeta(n, shape1 = p[["alpha"]], shape2 = p[["beta"]])
+           },
+           "exponential" = {
+               lambda <- x$parameters[["lambda"]]
+               stats::rexp(n = n, rate = lambda)
+           },
+           "gamma" = {
+               alpha <- x$parameters[["alpha"]]
+               beta <- x$parameters[["beta"]]
+               stats::rgamma(n = n, shape = alpha, rate = beta)
            },
            stop("Unknown prior type."))
 }