+++ R/ui_projection-simulation.R
@@ -275,7 +275,7 @@ sample_from <- function(nm, at, dt = NULL, grid_size = NULL, end = NULL, error.d
 #'
 #' @examples
 #' m <- aquarium_mod
-#' m <- set_prior(m, constant(0), "lambda")
+#' m <- set_prior(m, constant_p(0), "lambda")
 #' m <- set_params(m, sample_params(m))
 #' proj <- project(m, end = 40)
 #' plot(proj)