+++ tests/testthat/test-ui_plotting.R
@@ -24,6 +24,7 @@ test_that("Basic plotting does not crash", {
                      "lambda_NH4" = 0, "upsilon_algae_to_daphnia" = 0.15, 
                      "upsilon_NH4_to_algae" = 0.25, "upsilon_daphnia_to_NH4" = 0.04,
                      "zeta" = 0.1)) %>%
+        set_priors(normal_p(0, 4), "", quiet = TRUE) %>%
         project(end = 10)
     z <- sample_from(x, at = c(0, 1, 1.5, 2, 2.5, 3))
     for (c in c("size", "prop")) {
@@ -49,8 +50,8 @@ test_that("Basic plotting does not crash", {
 
 test_that("ggtopo() does not crash", {
     if (requireNamespace("ggraph")) {
-        expect_error(ggtopo(aquarium_mod, edge = "line"), NA)
-        expect_error(ggtopo(trini_mod), NA)
-        expect_error(ggtopo(trini_mod, layout = "sugiyama"), NA)
+        expect_error(print(ggtopo(aquarium_mod, edge = "line")), NA)
+        expect_error(print(ggtopo(trini_mod)), NA)
+        expect_error(print(ggtopo(trini_mod, layout = "sugiyama")), NA)
     }
 })